Panic commit

+++ Source/Events/EventLogSequenceNumber.cs
@@ -8,7 +8,7 @@ namespace Dolittle.Events
     /// <summary>
     /// Represents the sequence number of the Event Log as a natural number, corresponding to the number of events that has been committed to the Event Store.
     /// </summary>
-    public class EventLogSequenceNumber : ConceptAs<uint>
+    public class EventLogSequenceNumber : ConceptAs<ulong>
     {
         /// <summary>
         /// The initial version of the Event Store before any Events are committed.
@@ -19,6 +19,6 @@ namespace Dolittle.Events
         /// Implicitly convert a <see cref="uint"/> to an <see cref="EventLogSequenceNumber"/>.
         /// </summary>
         /// <param name="number">The number.</param>
-        public static implicit operator EventLogSequenceNumber(uint number) => new EventLogSequenceNumber { Value = number };
+        public static implicit operator EventLogSequenceNumber(ulong number) => new EventLogSequenceNumber { Value = number };
     }
 }
